server/requirements.txt
appdirs==1.4.3
asn1crypto==0.23.0
autopep8==1.3.3
awscli==1.11.172
beautifulsoup4==4.6.0
bleach==2.1.1
boto3==1.4.7
botocore==1.7.30
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
cryptography==2.1.1
cycler==0.10.0
decorator==4.1.2
docutils==0.14
entrypoints==0.2.3
evdev==0.7.0
flake8==3.4.1
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
isc==2.0
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lutris==0.4.13
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.4.3
notebook==5.2.0
numpy==1.13.3
packaging==16.8
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pusher==1.7.2
pyasn1==0.3.7
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
pyOpenSSL==17.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyusb==1.0.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
simplegeneric==0.8.1
six==1.11.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.3
